ha-to-o365-sync/requirements.txt
O365==2.1.7
PyYAML>=6.0
requests>=2.31.0
